Arrow key navigation to highlight & provisionally go to an artboard while dialog is still open. (Canceling dialog with Esc reverts to original view).
If filtering results in only 1 artboard in list, we select & go to that
artboard immediately without any arrow keys. If multiple hits, we don't move
the view immediately but pressing Enter will still go to the top hit - so
we semi-highlight the top hit to indicate this.

Populate list view immediately on open, which allows for pure arrow-key
nav with no typing in simple documents.